Modualized database functions.
