+++ bloodedguild_forums_api/api.py
@@ -10,6 +10,7 @@ import hashlib
 import psycopg2
 from psycopg2.extras import DictCursor
 from bloodedguild_forums_api.clean_html import clean_html
+from bloodedguild_forums_api.db import DatabaseAuth, DatabaseConnector, DatabaseStringConstructor
 testing = False
 if(not testing):
     from bloodedguild_forums_api.config import config
@@ -24,6 +25,12 @@ app = Flask(__name__)
 app.config['SECRET_KEY'] = config["secret"]
 app.config['JWT_EXPIRATION_DELTA'] = timedelta(weeks=1)
 api = Api(app)
+
+DatabaseConnector.db_string_constructor = DatabaseStringConstructor(
+    config["db_name"],
+    config["username"],
+    config["password"]
+)
 
 def get_post_count(thread_id):
     db = DatabaseConnector()
@@ -65,7 +72,6 @@ def identity(payload):
 
 jwt = JWT(app, authenticate, identity)
 
-
 class ValidatorResource(Resource):
     def __init__(self, required_fields=[]):
         super().__init__()
@@ -84,69 +90,6 @@ class ValidatorResource(Resource):
                 abort(400, message=error_string)
         else:
             return json_data
-
-class DatabaseStringConstructor():
-    def __init__(self, db_name, user, password):
-        self.db_name = db_name
-        self.user = user
-        self.password = password
-
-    def __str__(self):
-        conn_string = "host='localhost' "
-        conn_string +="dbname='"+self.db_name+"' "
-        conn_string +="user='"+self.user+"' "
-        conn_string +="password='"+self.password+"' "
-        return conn_string
-
-
-class DatabaseConnector():
-    def __init__(self):
-        self.db_string_constructor = DatabaseStringConstructor(
-            DB_NAME,
-            USER,
-            PASSWORD
-        )
-        self.open()
-
-    def get_conn_string(self):
-        return str(self.db_string_constructor)
-
-    def get_cursor(self):
-        return self.connection.cursor()
-
-    def rollback(self):
-        self.connection.rollback()
-        self.connection.close()
-
-    def close(self):
-        self.connection.commit()
-        self.connection.close()
-
-    def open(self):
-        self.connection = psycopg2.connect(
-            self.get_conn_string(),
-            cursor_factory=DictCursor
-        )
-
-
-class DatabaseAuth():
-    def __init__(self):
-        self.db = DatabaseConnector()
-
-    def check_auth(self, username, password):
-        psql_cursor = self.db.get_cursor()
-        sql_string = "select id, username, name from users_post_counts "
-        sql_string += "where username = %s"
-        sql_string += "and password_hash = crypt(%s, password_hash);"
-        psql_cursor.execute(
-            sql_string,
-            (username, password)
-        )
-        query_result = psql_cursor.fetchone()
-        psql_cursor.close()
-        self.db.close()
-        return query_result
-
 
 class DefaultLocation(Resource):
     def get(self):
